some changes to css
