+++ app/package.json
@@ -52,6 +52,11 @@
   },
   "devDependencies": {
     "@types/js-cookie": "^3.0.2",
-    "typescript-plugin-css-modules": "^3.4.0"
+    "html-webpack-plugin": "^5.5.0",
+    "style-loader": "^3.3.1",
+    "typescript-plugin-css-modules": "^3.4.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.0"
   }
 }
