+++ app/src/components/Profile/Profile.module.css
@@ -1,8 +1,8 @@
-* {
+/* * {
     box-sizing: border-box;
     margin: 0;
     padding:0;
-}
+} */
 
 .profile-container {
     border-style: solid;
@@ -81,7 +81,6 @@
     display: flex;
     flex-direction: column;
     min-height: 100%;
-    
 }
 
 .primary-stats h3 {
